Add product type endpoints; take out resolvers to separate file
